src/data_ingestion/redtrack/requirements.txt
requests==2.28.2
python-dotenv==1.0.0
psycopg2-binary==2.9.5
SQLAlchemy==2.0.4
pandas==1.5.3
schedule==1.1.0
